requirements.txt
fastapi~=0.111.1
pydantic~=2.8.2
uvicorn~=0.30.1
PyJWT~=2.8.0
python-decouple~=3.8
psycopg2-binary~=2.9.9
python-dotenv~=1.0.1
bcrypt~=4.1.3
openai~=1.37.0
boto3~=1.35.3
botocore~=1.35.3
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1